rest

+++ SpringMVC/src/main/webapp/user/js/user.js
@@ -57,6 +57,26 @@ const handleUpdateUser = (id) => { console.log('按下修改:' + id); };
 
 const handleDeleteUser = async(id) => { 
 	console.log('按下刪除:' + id); 
+	// 確認是否要刪除 ?
+	/*
+	if(!confirm('是否要刪除?')) {
+		return;
+	}
+	*/
+	const result = await Swal.fire({
+		title: '確定要刪除嗎?',
+		text: '刪除後將無法恢復',
+		icon: 'warning',
+		showCancelButton: true,
+		confirmButtonText: '是的, 刪除它!',
+		cancelButtonText: '取消'
+	});
+	
+	if(!result.ifConfirmed) {
+		return;
+	}
+	
+	// 進行刪除程序
 	const fullUrl = 'http://localhost:8080/SpringMVC/mvc/rest/user/' + id;
 	const response = await fetch(fullUrl, {method: 'DELETE'}); // 等待 fetch 請求完成
 	const {state, message, data} = await response.json(); // 等待回應本文內容
